Update Team_Spirit_Detail_Design.md
Added changes:
Added conditions in constraint solution 1 and 8. stipulating that the front border must be painted with university colors to abide by constraints 1 and 8.

Added a relay module to relay the DIO signals of the microcontroller with power provided by the power system. 

Addressing Feedback: 

The system was not previously capable of powering the LEDs however with the new changes to the system, the LEDs will be powered with no predicted issues. 

The dynamic display of team spirit must not be on the front of the robot however it must be identifiable from all sides. The system is still dynamic however it is reasonable that a static component of team spirit may be added to the front of the robot in the form of a gold stripe above a purple stripe. This prevents tampering with the navigation system and creates a display from all sides of the robot.

Remaining changes:

Development of a new KiCAD model which includes the 2-channel relay module.