Update Team_Spirit_Detail_Design.md
Added a 9th constraint to account for the lifespan of the relay. it was determined that the minimum lifespan of the relay is 10,000 operations, or 41 minutes high before the relay will potentially break. The solution outlines the budgeting of time for debugging and the competition to determine that the relay is in fact an acceptable component for this project and will not significantly increase the risk of failure.